Moved JEI support to /apihandlers
