+++ gradle.properties
@@ -3,7 +3,7 @@
 org.gradle.jvmargs=-Xmx3G
 
 mc_version=1.12.2
-jei_version=4.9.0.163
+jei_version=4.9.1.179
 
 mod_version=1.3.1
 forge_version=1.12.2-14.23.3.2655
