+++ src/main/java/notreepunching/apihandlers/jei/knife/KnifeRecipeCategory.java
@@ -1,4 +1,4 @@
-package notreepunching.jei.knife;
+package notreepunching.apihandlers.jei.knife;
 
 import mezz.jei.api.IGuiHelper;
 import mezz.jei.api.gui.IDrawable;
@@ -6,11 +6,9 @@ import mezz.jei.api.gui.IRecipeLayout;
 import mezz.jei.api.ingredients.IIngredients;
 import mezz.jei.api.recipe.IRecipeCategory;
 import mezz.jei.api.recipe.IRecipeWrapper;
-import net.minecraft.client.Minecraft;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 import notreepunching.NoTreePunching;
-import notreepunching.item.ModItems;
 
 import javax.annotation.Nonnull;
